added icon button active css, renamed element
