+++ lib/ui/elements/_elements.mjs
@@ -12,7 +12,7 @@ import dropdown from './dropdown.mjs'
 
 import dropdown_multi from './dropdown_multi.mjs'
 
-import iconButton from './iconButton.mjs'
+import buttonIcon from './buttonIcon.mjs'
 
 import isoline_params_here from './isoline_params_here.mjs'
 
@@ -36,7 +36,7 @@ export default {
   drawing,
   dropdown,
   dropdown_multi,
-  iconButton,
+  buttonIcon,
   isoline_params_here,
   isoline_params_mapbox,
   legendIcon,
